third_party/blink/web_tests/fast/media/mq-hover-expected.txt

Test the (hover) media features. See Bug 136119 for details.

------------- with 'hover' hover, 'hover' is primary -------------
Query "(hover)": true
Query "(Hover)": true
Query "(hover:none)": false
Query "(hover:NoNe)": false
Query "(hover:on-demand)": false
Query "(hover:bogusvalue)": false
Query "(hover:hover)": true
Query "(hover:HoveR)": true
Query "(any-hover)": true
Query "(any-Hover)": true
Query "(any-hover:none)": false
Query "(any-hover:NoNe)": false
Query "(any-hover:on-demand)": false
Query "(any-hover:bogusvalue)": false
Query "(any-hover:hover)": true
Query "(any-hover:HoveR)": true
------------- with 'none' and 'hover' hover, 'none' is primary -------------
Query "(hover)": false
Query "(Hover)": false
Query "(hover:none)": true
Query "(hover:NoNe)": true
Query "(hover:on-demand)": false
Query "(hover:bogusvalue)": false
Query "(hover:hover)": false
Query "(hover:HoveR)": false
Query "(any-hover)": true
Query "(any-Hover)": true
Query "(any-hover:none)": true
Query "(any-hover:NoNe)": true
Query "(any-hover:on-demand)": false
Query "(any-hover:bogusvalue)": false
Query "(any-hover:hover)": true
Query "(any-hover:HoveR)": true
------------- with 'none' hover, 'none' is primary -------------
Query "(hover)": false
Query "(Hover)": false
Query "(hover:none)": true
Query "(hover:NoNe)": true
Query "(hover:on-demand)": false
Query "(hover:bogusvalue)": false
Query "(hover:hover)": false
Query "(hover:HoveR)": false
Query "(any-hover)": false
Query "(any-Hover)": false
Query "(any-hover:none)": true
Query "(any-hover:NoNe)": true
Query "(any-hover:on-demand)": false
Query "(any-hover:bogusvalue)": false
Query "(any-hover:hover)": false
Query "(any-hover:HoveR)": false
